fix(PR-1877): colocate seat mutation guard tests
